Third milestone wip.

+++ src/sql/init.sql
@@ -1,5 +1,7 @@
 -- CREATE DATABASE ML_TESTING;
 
+
+
 -- DROP SCHEMA CORE CASCADE;
 -- DROP SCHEMA ADMIN CASCADE;
 -- DROP SCHEMA ANALYTICS CASCADE;
@@ -25,15 +27,15 @@ CREATE TABLE ADMIN.ROLE (
 
 CREATE TABLE ADMIN.USER (
     id serial primary key,
-    role_id serial references ADMIN.ROLE(id),
+    role_id smallint references ADMIN.ROLE(id),
     name varchar(20),
     surname varchar(80) not null,
     hourly_rate money
 );
 
 CREATE TABLE ADMIN.USER_LANG (
-    user_id serial references ADMIN.USER(id),
-    lang_id serial references ADMIN.LANG(id),
+    user_id smallint references ADMIN.USER(id),
+    lang_id smallint references ADMIN.LANG(id),
     proficiency_level smallint,
 
     primary key (user_id, lang_id)
@@ -56,8 +58,8 @@ CREATE TABLE CORE.MODEL_TYPE (
 
 CREATE TABLE CORE.MODEL (
     id serial primary key,
-    product_id serial references CORE.PRODUCT(id),
-    type_id serial references CORE.MODEL_TYPE(id),
+    product_id smallint references CORE.PRODUCT(id),
+    type_id smallint references CORE.MODEL_TYPE(id),
     name varchar(100) not null,
     major_version smallint not null,
     minor_version smallint not null,
@@ -73,23 +75,23 @@ CREATE TABLE CORE.COMPONENT (
 
 CREATE TABLE CORE.MODEL_COMPONENT (
     id serial primary key,
-    model_id serial references CORE.MODEL(id),
-    component_id serial references CORE.COMPONENT(id),
+    model_id smallint references CORE.MODEL(id),
+    component_id smallint references CORE.COMPONENT(id),
     order_number smallint not null
 );
 
 CREATE TABLE CORE.EVALUATION (
     id serial primary key,
-    model_id serial references CORE.MODEL(id),
-    user_id serial references ADMIN.USER(id),
+    model_id smallint references CORE.MODEL(id),
+    user_id smallint references ADMIN.USER(id),
     report varchar(500),
     is_completed_flag bit not null
 );
 
 CREATE TABLE CORE.CORPUS (
     id serial primary key,
-    domain_id serial references CORE.DOMAIN(id),
-    lang_id serial references ADMIN.LANG(id),
+    domain_id smallint references CORE.DOMAIN(id),
+    lang_id smallint references ADMIN.LANG(id),
     name varchar(100),
     major_version smallint,
     minor_version smallint,
@@ -104,33 +106,33 @@ CREATE TABLE CORE.ROW (
 );
 
 CREATE TABLE CORE.CORPUS_ROW (
-    corpus_id serial references CORE.CORPUS(id),
-    row_id serial references CORE.ROW(id),
+    corpus_id smallint references CORE.CORPUS(id),
+    row_id smallint references CORE.ROW(id),
     primary key (corpus_id, row_id)
 );
 
 CREATE TABLE CORE.HYPOTHESIS (
     id serial primary key,
-    row_id serial references CORE.ROW(id),
-    model_component_id serial references CORE.MODEL_COMPONENT(id),
+    row_id smallint references CORE.ROW(id),
+    model_component_id smallint references CORE.MODEL_COMPONENT(id),
     text varchar(200) not null,
     score real not null
 );
 
 CREATE TABLE CORE.COMMENT (
     id serial primary key,
-    hypothesis_id int references core.hypothesis(id),
-    evaluation_id int references core.evaluation(id),
+    hypothesis_id smallint references core.hypothesis(id),
+    evaluation_id smallint references core.evaluation(id),
     comment varchar(200)
 );
 
 CREATE TABLE ANALYTICS.WORK_TIME (
     id serial primary key,
-    user_id serial references ADMIN.USER(id),
+    user_id smallint references ADMIN.USER(id),
     start_time timestamp not null,
     finish_time timestamp
 );
-
+--
 -- CREATE USER ml_db_admin;
 --
 -- GRANT pg_read_all_data TO ml_db_admin;
